updated
Traefik v2 had some issues with VPN and I haven't been bothered with converting it from v1 to v2, forcing v1 now.

Removed Grafana and everything related to that as it was unneccesary.

Disabled Ouroboros as it is not compatible with this VPN setup right now. Issue created in their github project.

Added Sonarr 4K and Radarr 4K for glorious UHD resolution.